requirements.txt
asgiref==3.4.1
boto3==1.18.45
botocore==1.21.45
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.41.0
django-annoying==0.10.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.3.2
psycopg2-binary==2.9.1
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
sqlparse==0.4.1
stripe==2.60.0
